feat: Refactored builder component
- Refactored Builder component
- Replaced outdated Enzyme with RTC
- Added possibility to customize HistoryControls
- Updated Documentation and API on the library website
