+++ package.json
@@ -64,9 +64,6 @@
   "publishConfig": {
     "registry": "https://registry.npmjs.org"
   },
-  "overrides": {
-    "cheerio": "1.0.0-rc.12"
-  },
   "files": [
     "dist"
   ],
@@ -100,15 +97,14 @@
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@cfaester/enzyme-adapter-react-18": "^0.8.0",
     "@eslint/js": "^9.15.0",
-    "@types/enzyme": "^3.10.19",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
+    "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^29.5.14",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/styled-components": "^5.1.34",
-    "enzyme": "^3.11.0",
-    "enzyme-to-json": "^3.6.2",
     "eslint": "^9.15.0",
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-react-hooks": "^5.0.0",
