+++ setup-tests.ts
@@ -1,19 +1,6 @@
-import { configure } from 'enzyme';
-import Adapter from '@cfaester/enzyme-adapter-react-18';
+import '@testing-library/jest-dom';
+import { cleanup } from '@testing-library/react';
 
-configure({ adapter: new Adapter() });
-
-const originalConsoleError = console.error;
-
-console.error = (...args: unknown[]) => {
-  const [firstArg] = args;
-
-  if (
-    typeof firstArg === 'string' &&
-    firstArg.includes('findDOMNode is deprecated')
-  ) {
-    return;
-  }
-
-  originalConsoleError(...args);
-};
+afterEach(() => {
+  cleanup();
+});
