+++ src/builder/components/root-controls.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const RootControls = styled.div`
+  display: grid;
+  grid-auto-columns: min-content;
+  grid-auto-flow: column;
+  grid-gap: 0.5rem;
+  justify-content: start;
+  margin-bottom: 0.5rem;
+`;
