+++ src/widgets/field-select.test.tsx
@@ -1,5 +1,5 @@
-import { mount, shallow } from 'enzyme';
-import React from 'react';
+import React, { ReactElement } from 'react';
+import { fireEvent, render } from '@testing-library/react';
 import {
   IBuilderComponentsProps,
   IBuilderFieldProps,
@@ -10,47 +10,13 @@ import { FieldSelect } from './field-select';
 
 const components: IBuilderComponentsProps = defaultComponents;
 const fields: IBuilderFieldProps[] = [
-  {
-    field: 'MOCK_FIELD_1',
-    label: 'Mock Field',
-    type: 'BOOLEAN',
-  },
-  {
-    field: 'MOCK_FIELD_2',
-    label: 'Mock Field',
-    type: 'DATE',
-    operators: ['BETWEEN'],
-  },
-  {
-    field: 'MOCK_FIELD_3',
-    label: 'Mock Field',
-    type: 'DATE',
-    operators: ['LARGER'],
-  },
-  {
-    field: 'MOCK_FIELD_4',
-    label: 'Mock Field',
-    type: 'TEXT',
-    operators: ['EQUAL'],
-  },
-  {
-    field: 'MOCK_FIELD_5',
-    label: 'Mock Field',
-    type: 'TEXT',
-    operators: ['BETWEEN'],
-  },
-  {
-    field: 'MOCK_FIELD_6',
-    label: 'Mock Field',
-    type: 'NUMBER',
-    operators: ['EQUAL'],
-  },
-  {
-    field: 'MOCK_FIELD_7',
-    label: 'Mock Field',
-    type: 'NUMBER',
-    operators: ['BETWEEN'],
-  },
+  { field: 'MOCK_FIELD_1', label: 'Mock Field', type: 'BOOLEAN' },
+  { field: 'MOCK_FIELD_2', label: 'Mock Field', type: 'DATE', operators: ['BETWEEN'] },
+  { field: 'MOCK_FIELD_3', label: 'Mock Field', type: 'DATE', operators: ['LARGER'] },
+  { field: 'MOCK_FIELD_4', label: 'Mock Field', type: 'TEXT', operators: ['EQUAL'] },
+  { field: 'MOCK_FIELD_5', label: 'Mock Field', type: 'TEXT', operators: ['BETWEEN'] },
+  { field: 'MOCK_FIELD_6', label: 'Mock Field', type: 'NUMBER', operators: ['EQUAL'] },
+  { field: 'MOCK_FIELD_7', label: 'Mock Field', type: 'NUMBER', operators: ['BETWEEN'] },
   {
     field: 'MOCK_FIELD_8',
     label: 'Mock Field',
@@ -70,99 +36,69 @@ const fields: IBuilderFieldProps[] = [
     value: 'test',
   },
 ];
-const data: any[] = [
-  {
-    id: 'test',
-    field: 'MOCK_FIELD_1',
-    value: false,
-  },
-];
-const strings = { form: {} };
-const setData = jest.fn();
-const onChange = () => jest.fn();
+const data: any[] = [{ id: 'test', field: 'MOCK_FIELD_1', value: false }];
+
+const getByDataTest = (container: HTMLElement, value: string): HTMLElement => {
+  const element = container.querySelector(`[data-test="${value}"]`);
+
+  if (!element) {
+    throw new Error(`Unable to find element with data-test="${value}"`);
+  }
+
+  return element as HTMLElement;
+};
+
+const renderWithContext = (
+  element: ReactElement,
+  overrides?: Partial<React.ComponentProps<typeof BuilderContext.Provider>['value']>
+) =>
+  render(
+    <BuilderContext.Provider
+      value={{
+        components,
+        fields,
+        data,
+        strings: { form: {} },
+        setData: jest.fn(),
+        onChange: jest.fn(),
+        readOnly: false,
+        ...overrides,
+      }}
+    >
+      {element}
+    </BuilderContext.Provider>
+  );
 
 describe('#components/Widgets/FieldSelect', () => {
-  it('Tests Snapshot', () => {
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{
-            components,
-            fields,
-            data,
-            strings,
-            setData,
-            onChange,
-            readOnly: false,
-          }}
-        >
-          <FieldSelect id="test" selectedValue={''} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{
-            components,
-            fields,
-            data,
-            strings,
-            setData,
-            onChange,
-            readOnly: true,
-          }}
-        >
-          <FieldSelect id="test" selectedValue={''} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-  });
-
-  it('Tests user interaction', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{
-          components,
-          fields,
-          data,
-          strings,
-          setData,
-          onChange,
-          readOnly: false,
-        }}
-      >
-        <FieldSelect id="test" selectedValue={''} />
-      </BuilderContext.Provider>
-    );
-
-    fields.forEach(item => {
-      wrapper.find('[data-test="SelectMultiTrigger"]').first().simulate('click');
-      wrapper
-        .find(`[data-test="SelectMultiOption[${item.field}]"]`)
-        .first()
-        .simulate('click');
+  it('renders in editable and read-only modes', () => {
+    const editable = renderWithContext(<FieldSelect id="test" selectedValue="" />);
+    const readOnly = renderWithContext(<FieldSelect id="test" selectedValue="" />, {
+      readOnly: true,
     });
+
+    expect(editable.container.firstChild).toBeTruthy();
+    expect(readOnly.container.firstChild).toBeTruthy();
   });
 
-  it('Test no form components scenario', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{
-          components: {},
-          fields,
-          data,
-          strings,
-          setData,
-          onChange,
-          readOnly: false,
-        }}
-      >
-        <FieldSelect id="test" selectedValue={''} />
-      </BuilderContext.Provider>
+  it('cycles through the available fields', () => {
+    const { container } = renderWithContext(
+      <FieldSelect id="test" selectedValue="" />
     );
 
-    expect(Object.keys(wrapper).length).toBe(0);
+    for (const item of fields) {
+      fireEvent.click(getByDataTest(container, 'SelectMultiTrigger'));
+      fireEvent.click(getByDataTest(container, `SelectMultiOption[${item.field}]`));
+    }
+
+    expect(container.firstChild).toBeTruthy();
+  });
+
+  it('falls back to the default form components when custom ones are unavailable', () => {
+    const { container } = renderWithContext(
+      <FieldSelect id="test" selectedValue="" />,
+      { components: {} as IBuilderComponentsProps }
+    );
+
+    expect(container.querySelector('[data-test="SelectMultiTrigger"]')).toBeTruthy();
   });
 });
