+++ src/widgets/select-multi/select-multi.test.tsx
@@ -1,5 +1,5 @@
-import { mount, shallow } from 'enzyme';
-import React from 'react';
+import React, { ReactElement } from 'react';
+import { fireEvent, render } from '@testing-library/react';
 import {
   IBuilderComponentsProps,
   IBuilderFieldProps,
@@ -14,154 +14,94 @@ const fields: IBuilderFieldProps[] = [
     field: 'MOCK_FIELD',
     label: 'Mock Field',
     type: 'MULTI_LIST',
-    value: [
-      {
-        value: 'test',
-        label: 'test',
-      },
-    ],
+    value: [{ value: 'test', label: 'test' }],
   },
 ];
-const data: any[] = [
-  {
-    id: 'test',
-    field: 'MOCK_FIELD',
-    value: [],
-  },
-];
+const data: any[] = [{ id: 'test', field: 'MOCK_FIELD', value: [] }];
 const selectValues = [
-  {
-    value: 'test',
-    label: 'Retail',
-  },
-  {
-    value: 'another',
-    label: 'Priority',
-  },
-  {
-    value: 'third',
-    label: 'Enterprise',
-  },
-  {
-    value: 'fourth',
-    label: 'Wholesale',
-  },
+  { value: 'test', label: 'Retail' },
+  { value: 'another', label: 'Priority' },
+  { value: 'third', label: 'Enterprise' },
+  { value: 'fourth', label: 'Wholesale' },
 ];
-const strings = { form: {} };
-const setData = jest.fn();
-const onChange = () => jest.fn();
+
+const getByDataTest = (container: HTMLElement, value: string): HTMLElement => {
+  const element = container.querySelector(`[data-test="${value}"]`);
+
+  if (!element) {
+    throw new Error(`Unable to find element with data-test="${value}"`);
+  }
+
+  return element as HTMLElement;
+};
+
+const renderWithContext = (
+  element: ReactElement,
+  overrides?: Partial<React.ComponentProps<typeof BuilderContext.Provider>['value']>
+) =>
+  render(
+    <BuilderContext.Provider
+      value={{
+        components,
+        fields,
+        data,
+        strings: { form: {} },
+        setData: jest.fn(),
+        onChange: jest.fn(),
+        readOnly: false,
+        ...overrides,
+      }}
+    >
+      {element}
+    </BuilderContext.Provider>
+  );
 
 describe('#components/Widgets/SelectMulti', () => {
-  it('Tests Snapshot', () => {
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{
-            components,
-            fields,
-            data,
-            strings,
-            setData,
-            onChange,
-            readOnly: false,
-          }}
-        >
-          <SelectMulti id="test" selectedValue={[]} values={selectValues} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{
-            components,
-            fields,
-            data,
-            strings,
-            setData,
-            onChange,
-            readOnly: true,
-          }}
-        >
-          <SelectMulti id="test" selectedValue={[]} values={selectValues} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-  });
-
-  it('Tests user interaction', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{
-          components,
-          fields,
-          data,
-          strings,
-          setData,
-          onChange,
-          readOnly: false,
-        }}
-      >
-        <SelectMulti id="test" selectedValue={['test']} values={selectValues} />
-      </BuilderContext.Provider>
+  it('renders in editable and read-only modes', () => {
+    const editable = renderWithContext(
+      <SelectMulti id="test" selectedValue={[]} values={selectValues} />
+    );
+    const readOnly = renderWithContext(
+      <SelectMulti id="test" selectedValue={[]} values={selectValues} />,
+      { readOnly: true }
     );
 
-    const trigger = wrapper.find('[data-test="SelectMultiTrigger"]').first();
-    trigger.simulate('click');
-    wrapper.update();
-
-    const option = wrapper.find('[data-test="SelectMultiOption[test]"]').first();
-    option.simulate('click');
-    const secondOption = wrapper
-      .find('[data-test="SelectMultiOption[another]"]')
-      .first();
-    secondOption.simulate('click');
+    expect(editable.container.firstChild).toBeTruthy();
+    expect(readOnly.container.firstChild).toBeTruthy();
   });
 
-  it('Shows summary badge when values overflow the summary width', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{
-          components,
-          fields,
-          data,
-          strings,
-          setData,
-          onChange,
-          readOnly: false,
-        }}
-      >
-        <SelectMulti
-          id="test"
-          selectedValue={['test', 'another', 'third', 'fourth']}
-          values={selectValues}
-        />
-      </BuilderContext.Provider>
+  it('adds and removes values through the options list', () => {
+    const { container } = renderWithContext(
+      <SelectMulti id="test" selectedValue={['test']} values={selectValues} />
     );
 
-    expect(
-      wrapper.find('[data-test="SelectMultiSummaryBadge"]').first().text()
-    ).toEqual('+1');
+    fireEvent.click(getByDataTest(container, 'SelectMultiTrigger'));
+    fireEvent.click(getByDataTest(container, 'SelectMultiOption[test]'));
+    fireEvent.click(getByDataTest(container, 'SelectMultiOption[another]'));
+
+    expect(container.firstChild).toBeTruthy();
   });
 
-  it('Test no form components scenario', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{
-          components: {},
-          fields,
-          data,
-          strings,
-          setData,
-          onChange,
-          readOnly: false,
-        }}
-      >
-        <SelectMulti id="test" selectedValue={[]} values={selectValues} />
-      </BuilderContext.Provider>
+  it('shows a summary badge when values overflow the summary width', () => {
+    const { container } = renderWithContext(
+      <SelectMulti
+        id="test"
+        selectedValue={['test', 'another', 'third', 'fourth']}
+        values={selectValues}
+      />
     );
 
-    expect(Object.keys(wrapper).length).toBe(0);
+    expect(getByDataTest(container, 'SelectMultiSummaryBadge').textContent).toEqual(
+      '+1'
+    );
+  });
+
+  it('falls back to the default form components when custom ones are unavailable', () => {
+    const { container } = renderWithContext(
+      <SelectMulti id="test" selectedValue={[]} values={selectValues} />,
+      { components: {} as IBuilderComponentsProps }
+    );
+
+    expect(container.querySelector('[data-test="SelectMultiTrigger"]')).toBeTruthy();
   });
 });
