+++ src/widgets/select.test.tsx
@@ -1,5 +1,5 @@
-import { mount, shallow } from 'enzyme';
-import React from 'react';
+import React, { ReactElement } from 'react';
+import { fireEvent, render } from '@testing-library/react';
 import {
   IBuilderComponentsProps,
   IBuilderFieldProps,
@@ -14,76 +14,74 @@ const fields: IBuilderFieldProps[] = [
     field: 'MOCK_FIELD',
     label: 'Mock Field',
     type: 'LIST',
-    value: [
-      {
-        value: 'test',
-        label: 'test',
-      },
-    ],
+    value: [{ value: 'test', label: 'test' }],
   },
 ];
-const data: any[] = [
-  {
-    id: 'test',
-    field: 'MOCK_FIELD',
-    value: 'test',
-  },
-];
-const selectValues = [
-  {
-    value: 'test',
-    label: 'test',
-  },
-];
-const strings = { form: {} };
-const setData = jest.fn();
-const onChange = () => jest.fn();
+const data: any[] = [{ id: 'test', field: 'MOCK_FIELD', value: 'test' }];
+const selectValues = [{ value: 'test', label: 'test' }];
+
+const getByDataTest = (container: HTMLElement, value: string): HTMLElement => {
+  const element = container.querySelector(`[data-test="${value}"]`);
+
+  if (!element) {
+    throw new Error(`Unable to find element with data-test="${value}"`);
+  }
+
+  return element as HTMLElement;
+};
+
+const renderWithContext = (
+  element: ReactElement,
+  overrides?: Partial<React.ComponentProps<typeof BuilderContext.Provider>['value']>
+) =>
+  render(
+    <BuilderContext.Provider
+      value={{
+        components,
+        fields,
+        data,
+        strings: { form: {} },
+        setData: jest.fn(),
+        onChange: jest.fn(),
+        readOnly: false,
+        ...overrides,
+      }}
+    >
+      {element}
+    </BuilderContext.Provider>
+  );
 
 describe('#components/Widgets/Select', () => {
-  it('Tests Snapshot', () => {
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{ components, fields, data, strings, setData, onChange, readOnly: false }}
-        >
-          <Select id="test" selectedValue={''} values={selectValues} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-
-    expect(
-      shallow(
-        <BuilderContext.Provider
-          value={{ components, fields, data, strings, setData, onChange, readOnly: true }}
-        >
-          <Select id="test" selectedValue={''} values={selectValues} />
-        </BuilderContext.Provider>
-      )
-    ).toMatchSnapshot();
-  });
-
-  it('Tests user interaction', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{ components, fields, data, strings, setData, onChange, readOnly: false }}
-      >
-        <Select id="test" selectedValue={''} values={selectValues} />
-      </BuilderContext.Provider>
+  it('renders in editable and read-only modes', () => {
+    const editable = renderWithContext(
+      <Select id="test" selectedValue="" values={selectValues} />
+    );
+    const readOnly = renderWithContext(
+      <Select id="test" selectedValue="" values={selectValues} />,
+      { readOnly: true }
     );
 
-    wrapper.find('[data-test="SelectMultiTrigger"]').first().simulate('click');
-    wrapper.find('[data-test="SelectMultiOption[test]"]').first().simulate('click');
+    expect(editable.container.firstChild).toBeTruthy();
+    expect(readOnly.container.firstChild).toBeTruthy();
   });
 
-  it('Test no form components scenario', () => {
-    const wrapper = mount(
-      <BuilderContext.Provider
-        value={{ components: {}, fields, data, strings, setData, onChange, readOnly: false }}
-      >
-        <Select id="test" selectedValue={''} values={selectValues} />
-      </BuilderContext.Provider>
+  it('opens the options and selects a value', () => {
+    const { container } = renderWithContext(
+      <Select id="test" selectedValue="" values={selectValues} />
     );
 
-    expect(Object.keys(wrapper).length).toBe(0);
+    fireEvent.click(getByDataTest(container, 'SelectMultiTrigger'));
+    fireEvent.click(getByDataTest(container, 'SelectMultiOption[test]'));
+
+    expect(container.firstChild).toBeTruthy();
+  });
+
+  it('falls back to the default form components when custom ones are unavailable', () => {
+    const { container } = renderWithContext(
+      <Select id="test" selectedValue="" values={selectValues} />,
+      { components: {} as IBuilderComponentsProps }
+    );
+
+    expect(container.querySelector('[data-test="SelectMultiTrigger"]')).toBeTruthy();
   });
 });
